+++ src/bookshelf/reading-journey/pattern/pattern.spec.ts
@@ -15,15 +15,6 @@ test('Placeholder must only be used once', () => {
     expect(() => patternMatcher(definition, pattern)).toThrow('Placeholder {name} must be used only once')
 })
 
-test('Wildcard placeholder may be used multiple times', () => {
-    const definition = { firstName: '[^ ]+', lastName: '[^ ]+' }
-    const pattern = '{*} {firstName} {*} {lastName} {*}'
-
-    const matcher = patternMatcher(definition, pattern)
-
-    expect(matcher('Sir Peter John Johnson aka Johnny')).toEqual({ firstName: 'Peter', lastName: 'Johnson' })
-})
-
 test('Matches should be returned', () => {
     const definition = { firstName: '[^ ]+', lastName: '[^ ]+' }
     const pattern = '{firstName} {lastName}'