+++ src/bookshelf/reading-journey/pattern/pattern.ts
@@ -28,7 +28,7 @@ export function patternMatcher<T extends PlaceholderPatterns>(definition: T, pat
         placeholders.set(match[0], (placeholders.get(match[0]) || 0) + 1)
     }
 
-    let regexString = pattern.replace(/\{\*}/g, '__PLACEHOLDER_WILDCARD__')
+    let regexString = pattern
     for (const name of Object.keys(definition)) {
         const placeholder = `{${name}}`
         if (!placeholders.has(placeholder)) {
@@ -42,9 +42,7 @@ export function patternMatcher<T extends PlaceholderPatterns>(definition: T, pat
         regexString = regexString.replace(`{${name}}`, `__PLACEHOLDER__${name}__`)
     }
 
-    regexString = regexString
-        .replace(/[.*+?^${}()|[\]\\]/g, '\\$&') // @see https://stackoverflow.com/a/6969486
-        .replace(/__PLACEHOLDER_WILDCARD__/g, '.*?')
+    regexString = regexString.replace(/[.*+?^${}()|[\]\\]/g, '\\$&') // @see https://stackoverflow.com/a/6969486
 
     for (const [name, pattern] of Object.entries(definition)) {
         regexString = regexString.replace(`__PLACEHOLDER__${name}__`, `(?<${name}>${pattern})`)